% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{bibentry-package}
\alias{bibentry}
\alias{bibentry-package}
\title{bibentry: Advanced Manipulation of bibentry}
\description{
Advanced Manipulation of bibentry.
}
\examples{
(b1 = 'rmarkdown' |> citation())
(b2 = 'ggplot2' |> citation())

b1 |> sort_by(y = 'year', decreasing = TRUE)
b1 |> sort_by(y = 'bibtype')

b1 |> subset(subset = (bibtype != 'Manual'))
b1 |> subset(subset = (bibtype == 'Manual'))
b2 |> subset(subset = (bibtype == 'Book'))

}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

Authors:
\itemize{
  \item Tingting Zhan \email{tingtingzhan@gmail.com}
}

}
\keyword{internal}
