% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies2bib.R
\name{dependencies2bib}
\alias{dependencies2bib}
\title{Write All Denpendency's Citatation to \code{.bib} File}
\usage{
dependencies2bib(package, path = tempdir(), ...)
}
\arguments{
\item{package}{\link[base]{character} scalar}

\item{path}{\link[base]{character} scalar, directory of the output \code{.bib} file}

\item{...}{additional parameters, currently of no use}
}
\description{
Write All Denpendency's Citatation to \code{.bib} File
}
\examples{
\dontrun{
'nlme' |> dependencies2bib(path = tempdir()) # need to have repo to pass devtools::check
'spatstat' |> dependencies2bib(path = tempdir())
}
}
\keyword{internal}
