% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibs.R
\name{literature}
\alias{literature}
\alias{.fisher22}
\alias{.holm79}
\alias{.hochberg88}
\alias{.hommel88}
\alias{.benjamini_hochberg95}
\alias{.benjamini_yekutieli01}
\alias{.lawley_maxwell71}
\alias{.tukey49}
\title{Literature}
\usage{
.fisher22()

.holm79()

.hochberg88()

.hommel88()

.benjamini_hochberg95()

.benjamini_yekutieli01()

.lawley_maxwell71()

.tukey49()
}
\description{
Literature
}
\keyword{internal}
