% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package2key.R
\name{package2key}
\alias{package2key}
\title{Use Package as Key in \link[utils]{citation}}
\usage{
package2key(x)
}
\arguments{
\item{x}{a \link[utils]{citation} object}
}
\value{
Function \code{\link[=package2key]{package2key()}} returns a \link[utils]{citation} object.
}
\description{
Use Package as Key in \link[utils]{citation}
}
\examples{
citation() |> toBibtex()
citation() |> package2key() |> toBibtex()
'rmarkdown' |> citation() |> toBibtex()
'rmarkdown' |> citation() |> package2key() |> toBibtex()
}
\keyword{internal}
