% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{sink2}
\alias{sink2}
\alias{sink2.Bibtex}
\alias{sink2.bibentry}
\title{Sink an Object to a File}
\usage{
sink2(x, file, ...)

\method{sink2}{Bibtex}(x, file, ...)

\method{sink2}{bibentry}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{file, ...}{parameters of function \link[base]{sink}}
}
\description{
Sink an Object to a File
}
\note{
Too bad the function \link[base]{sink} is not an \code{S3} generic.
}
\keyword{internal}
