% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sink.R
\name{sink_bibentry}
\alias{sink_bibentry}
\title{Sink \link[utils]{bibentry} To a \code{.bib} File}
\usage{
sink_bibentry(x, file = "bibliography.bib", ...)
}
\arguments{
\item{x}{a \link[utils]{bibentry} object}

\item{file}{\link[base]{character} scalar, bibliography file name \code{'*.bib'}}

\item{...}{additional parameters, currently not in use}
}
\description{
Sink \link[utils]{bibentry} To a \code{.bib} File
}
\note{
Too bad the function \link[base]{sink} is not an \code{S3} generic.
}
\keyword{internal}
