% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{bibentry-package}
\alias{bibentry}
\alias{bibentry-package}
\title{bibentry: Advanced Manipulation of bibentry}
\description{
Advanced Manipulation of bibentry.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
